requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
boto3==1.38.13
botocore==1.38.13
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
fonttools==4.58.0
fsspec==2025.3.2
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.8.4
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.7
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pillow_heif==0.22.0
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pybboxes==0.1.6
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.5
roboflow==1.1.63
s3transfer==0.12.0
sahi==0.11.23
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.7.0
torchbnn==1.2
torchvision==0.22.0
tqdm==4.67.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.131
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
yolov5==7.0.14
pydantic-settings==2.9.1
